% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFormats.R
\docType{package}
\name{DEFormats}
\alias{DEFormats}
\alias{DEFormats-package}
\title{Convert between differential gene expression data formats}
\description{
\pkg{DEFormats} facilitates data conversion between various formats used by
different gene expression analysis packages.
}
\details{
Currently the package supports data conversion between \pkg{DESeq2} and 
\pkg{edgeR}, i.e. between the \code{\link[DESeq2]{DESeqDataSet-class}} and 
\code{\link[edgeR]{DGEList-class}}, objects, respectively.

Objects can be coerced using the following methods \itemize{ \item
\code{\link{as.DESeqDataSet}} \item \code{\link{as.DGEList}} }
}
\author{
Andrzej Oleś \email{andrzej.oles@embl.de}
}

