% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{DGEList,SummarizedExperiment0-method}
\alias{DGEList,SummarizedExperiment0-method}
\title{DGEList Constructor Generic}
\usage{
\S4method{DGEList}{SummarizedExperiment0}(counts = SummarizedExperiment(),
  lib.size = colSums(assay(counts)), norm.factors = rep(1, ncol(counts)),
  group = rep(1, ncol(counts)), genes = as.data.frame(rowRanges(counts)),
  remove.zeros = FALSE)
}
\arguments{
\item{counts}{numeric matrix of read counts}

\item{lib.size}{numeric vector giving the total count (sequence depth) for each library}

\item{norm.factors}{numeric vector of normalization factors that modify the library sizes}

\item{group}{vector or factor giving the experimental group/condition for each sample/library}

\item{genes}{data frame containing annotation information for each gene}

\item{remove.zeros}{logical, whether to remove rows that have 0 total count}
}
\description{
Create a \code{\link[edgeR]{DGEList-class}} object
}

