% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mockRnaSeqData}
\alias{mockRnaSeqData}
\title{Example counts table of RNA-seq data}
\format{A 1000 x 6 integer matrix with rows representing genes and columns different samples.}
\source{
The count table was generated using the \code{\link[DESeq2]{makeExampleDESeqDataSet}} method from the \pkg{DESeq2} package.
}
\usage{
data("mockRnaSeqData")
}
\description{
Simulated expression data of an RNA-seq experiment.
}
\examples{
data("mockRnaSeqData")
str(mockRnaSeqData)
}

