% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.DESeqDataSet}
\alias{as.DESeqDataSet}
\alias{as.DESeqDataSet.DGEList}
\title{Coerce objects to DESeqDataSet}
\usage{
as.DESeqDataSet(x, ...)

\method{as.DESeqDataSet}{DGEList}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
A \code{\link[DESeq2]{DESeqDataSet-class}} object
}
\description{
Coerce objects to \code{\link[DESeq2]{DESeqDataSet-class}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: Coerce \code{\link[edgeR]{DGEList-class}} objects to \code{\link[DESeq2]{DESeqDataSet-class}}.
}}
\examples{
require("edgeR")

counts = mockRnaSeqData()
group = rep(c("case", "control"), each = 3)

dge = DGEList(counts = counts, group = group)
dge

as.DESeqDataSet(dge)
}
\author{
Andrzej Oles <\email{andrzej.oles@embl.de}>
}

