% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.DESeqDataSet}
\alias{as.DESeqDataSet}
\alias{as.DESeqDataSet,DGEList-method}
\title{Coerce objects to DESeqDataSet}
\usage{
as.DESeqDataSet(x, ...)

\S4method{as.DESeqDataSet}{DGEList}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A \code{\link[DESeq2]{DESeqDataSet-class}} object
}
\description{
Coerce objects to \code{\link[DESeq2]{DESeqDataSet-class}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: Coerce \code{\link[edgeR]{DGEList-class}} objects to \code{\link[DESeq2]{DESeqDataSet-class}}.
}}
\examples{
require("edgeR")

data(mockRnaSeqData)
group = rep(c("case", "control"), each = 3)

dgelist <- DGEList(counts = mockRnaSeqData, group = group)
dgelist

as.DESeqDataSet(dgelist)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}
}

