% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.DGEList}
\alias{as.DGEList}
\alias{as.DGEList,DESeqDataSet-method}
\title{Coerce objects to DGEList}
\usage{
as.DGEList(x, ...)

\S4method{as.DGEList}{DESeqDataSet}(x)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
A \code{\link[edgeR]{DGEList-class}} object
}
\description{
Coerce objects to \code{\link[edgeR]{DGEList-class}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{DESeqDataSet}: Coerce \code{\link[DESeq2]{DESeqDataSet-class}} objects to \code{\link[edgeR]{DGEList-class}}.
}}
\examples{
require("DESeq2")

se = mockRnaSeqData(output = "SummarizedExperiment")
se

dds = DESeqDataSet(se, design = ~ condition)
dds

as.DGEList(dds)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}
}

