% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{DEFormats}
\alias{DEFormats}
\alias{DEFormats-package}
\title{Convert between various differential gene expression data formats}
\description{
\pkg{DEFormats} facilitates data conversion between different formats used by gene expression analysis packages.
}
\details{
Currently the package supports data conversion between the formats used by \pkg{DESeq2} and \pkg{edgeR}, i.e. \code{DESeqDataSet} and \code{DGEList}, respectively.

Objects can be coerced using the following methods
\code{as.DESeqDataSet}
\code{as.DGEList}
}

