% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{DEFormats}
\alias{DEFormats}
\alias{DEFormats-package}
\title{Convert between various differential gene expression data formats}
\description{
\pkg{DEFormats} facilitates data conversion between formats used by gene
expression analysis packages.
}
\details{
Currently the package supports data conversion between \pkg{DESeq2} and
\pkg{edgeR}, i.e. between the \code{\link[DESeq2]{DESeqDataSet-class}} and
\code{\link[edgeR]{DGEList-class}}, objects, respectively.

Objects can be coerced using the following methods
\itemize{
  \item \code{\link{as.DESeqDataSet}}
  \item \code{\link{as.DGEList}}
}
}

