% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{DGEList}
\alias{DGEList}
\alias{DGEList,RangedSummarizedExperiment-method}
\title{DGEList Constructor Generic}
\usage{
DGEList(counts = matrix(0, 0, 0), lib.size = colSums(counts),
  norm.factors = rep(1, ncol(counts)), samples = NULL, group = NULL,
  genes = NULL, remove.zeros = FALSE)


  \S4method{DGEList}{RangedSummarizedExperiment}(counts = new("RangedSummarizedExperiment"),
  lib.size = colData(counts)$lib.size,
  norm.factors = colData(counts)$norm.factors, samples = colData(counts),
  group = colData(counts)$group, genes = as.data.frame(rowRanges(counts)),
  remove.zeros = FALSE)
}
\arguments{
\item{counts}{read counts, either a numeric matrix or a
\linkS4class{RangedSummarizedExperiment} object}

\item{lib.size}{numeric vector giving the total count (sequence depth) for
each library}

\item{norm.factors}{numeric vector of normalization factors that modify the
library sizes}

\item{samples}{data frame containing information for each sample}

\item{group}{vector or factor giving the experimental group/condition for
each sample/library}

\item{genes}{data frame containing annotation information for each gene}

\item{remove.zeros}{logical, whether to remove rows that have 0 total count}
}
\value{
A \linkS4class{DGEList} object.
}
\description{
Creates a \linkS4class{DGEList} object.
}
\examples{
se = simulateRnaSeqData(output = "RangedSummarizedExperiment")

## Initialize a DGEList from a RangedSummarizedExperiment object
DGEList(se)
}
\author{
Andrzej Oleś <\email{andrzej.oles@embl.de}>, 2016
}

