% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockRnaSeqData.R
\name{mockRnaSeqData}
\alias{mockRnaSeqData}
\title{Example counts table of RNA-seq data}
\usage{
mockRnaSeqData(output = c("matrix", "SummarizedExperiment"), rep = 3,
  conditions = c("Case", "Control"), ...)
}
\arguments{
\item{output}{Output format}

\item{rep}{Number of replicates for each condition}

\item{conditions}{Condition names}

\item{...}{Arguments passed to \code{\link[DESeq2]{makeExampleDESeqDataSet}}}
}
\value{
Depending on the \code{output} setting a matrix or an \code{\link[GenomicRanges]{SummarizedExperiment-class}} object.
}
\description{
Simulated expression data of an RNA-seq experiment.
}
\details{
The count table is generated using the \code{\link[DESeq2]{makeExampleDESeqDataSet}} method from the \pkg{DESeq2} package.
}
\examples{
## count data
mockRnaSeqData = mockRnaSeqData()

## return an SummarizedExperiment object
mockRnaSeqDataSE = mockRnaSeqData(output = "SummarizedExperiment")

require("GenomicRanges")
identical(mockRnaSeqData, assay(mockRnaSeqDataSE))
}
\author{
Andrzej Oles <\email{andrzej.oles@embl.de}>
}

