% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRnaSeqData.R
\name{simulateRnaSeqData}
\alias{simulateRnaSeqData}
\title{Example counts table of RNA-seq data}
\usage{
simulateRnaSeqData(output = c("matrix", "RangedSummarizedExperiment"),
  rep = 3, conditions = c("Case", "Control"), seed = 0L, ...)
}
\arguments{
\item{output}{output type}

\item{rep}{number of replicates for each condition}

\item{conditions}{condition names}

\item{seed}{a single integer value specifying the random number generator seed}

\item{...}{arguments passed to \code{\link[DESeq2]{makeExampleDESeqDataSet}}}
}
\value{
Depending on the \code{output} setting a matrix or an \code{\linkS4class{RangedSummarizedExperiment}} object.
}
\description{
Simulated expression data of an RNA-seq experiment.
}
\details{
The count table is generated using the \code{\link[DESeq2]{makeExampleDESeqDataSet}} method from the \pkg{DESeq2} package.
}
\examples{
## count data
simulateRnaSeqData = simulateRnaSeqData()

## return an RangedSummarizedExperiment object
simulateRnaSeqDataSE = simulateRnaSeqData(output = "RangedSummarizedExperiment")

require("SummarizedExperiment")
identical(simulateRnaSeqData, assay(simulateRnaSeqDataSE))
}
\author{
Andrzej Oleś <\email{andrzej.oles@embl.de}>, 2016
}

